abstract estimateFee for evm
